adjusting structure
